Alot happened in these changes:
I added the StateManager as a service and have game logic. For now, I have implemented a test state, just to display stuff.

Also there is the Generic RCGameState class, modify it when you want to put logic abstract to all game states.

Also, I have constructed the cube graph structure. Just create an instance of RCCube and the whole graph will be created in the constructor.

Can be created for all odd cube sizes, will update to all sizes, even or odd.