Here it is... I have text implemented and boy was it a pain!
Text is realized through the text class, it comes from pane as each character is its own quad. The text can be oriented any way in 3d space.

We now have a font manager. This manages all the BitmapFonts witch have to be created uisng the tool I submitted with the tools. Then the application can load a font and use the font manager to access it anywhere. the font manager is a game service.

There are a couple of steps that need to be taken after the font has been produced. i will write a quick wiki on the code site. But call me for questions if you need to make another font.